help

+++ StrategyUnits/IHealthController.cs
@@ -5,7 +5,7 @@
         public int Health { get; set; }
         public int MaxHealth { get; protected set; }
         public bool IsAlive { get => Health > 0; }
-        public void TakeDamage(int damage);
+        public void TakeDamage(double damage);
         public void TakeHeal(int healAmount);
     }
 }
